Add License
Hello relaxbro,
we would like to use some source code of your project. But there is no license file and so we can't use it. You have also no e-mail address in your GitHub account, so that I can't get in contact with you. Thus I add a license file (MIT) and the year of your last commit (2017) and as copyright your GitHub account (relaxbro). Feel free the change it, but a GitHub project without license isn't useful for everyone because nobody can copy the code without a license.

Thanks in advance

    Klaus